WIP refactoring. NOT WORKING YET
