+++ impl11/ddraw/DeviceResources.h
@@ -7,163 +7,9 @@
 #include "../shaders/material_defs.h"
 #include "../shaders/shader_common.h"
 #include <vector>
-
-// Also found in the Floating_GUI_RESNAME list:
-extern const char *DC_TARGET_COMP_SRC_RESNAME;
-extern const char *DC_LEFT_SENSOR_SRC_RESNAME;
-extern const char *DC_LEFT_SENSOR_2_SRC_RESNAME;
-extern const char *DC_RIGHT_SENSOR_SRC_RESNAME;
-extern const char *DC_RIGHT_SENSOR_2_SRC_RESNAME;
-extern const char *DC_SHIELDS_SRC_RESNAME;
-extern const char *DC_SOLID_MSG_SRC_RESNAME;
-extern const char *DC_BORDER_MSG_SRC_RESNAME;
-extern const char *DC_LASER_BOX_SRC_RESNAME;
-extern const char *DC_ION_BOX_SRC_RESNAME;
-extern const char *DC_BEAM_BOX_SRC_RESNAME;
-extern const char *DC_TOP_LEFT_SRC_RESNAME;
-extern const char *DC_TOP_RIGHT_SRC_RESNAME;
-
-// Use the following with `const auto missionIndexLoaded = (int *)0x9F5E74;` to detect the DSII tunnel run mission.
-const int DEATH_STAR_MISSION_INDEX = 52;
-
-typedef struct Box_struct {
-	float x0, y0, x1, y1;
-} Box;
-
-typedef struct uvfloat4_struct {
-	float x0, y0, x1, y1;
-} uvfloat4;
-
-typedef struct float3_struct {
-	float x, y, z;
-} float3;
-
-typedef struct float4_struct {
-	float x, y, z, w;
-} float4;
-
-// Region names. Used in the erase_region and move_region commands
-const int LEFT_RADAR_HUD_BOX_IDX	= 0;
-const int RIGHT_RADAR_HUD_BOX_IDX	= 1;
-const int SHIELDS_HUD_BOX_IDX		= 2;
-const int BEAM_HUD_BOX_IDX			= 3;
-const int TARGET_HUD_BOX_IDX		= 4;
-const int LEFT_MSG_HUD_BOX_IDX		= 5;
-const int RIGHT_MSG_HUD_BOX_IDX		= 6;
-const int TOP_LEFT_HUD_BOX_IDX		= 7;
-const int TOP_RIGHT_HUD_BOX_IDX		= 8;
-const int TEXT_RADIOSYS_HUD_BOX_IDX	= 9;
-const int TEXT_CMD_HUD_BOX_IDX		= 10;
-const int MAX_HUD_BOXES				= 11;
-extern std::vector<const char *>g_HUDRegionNames;
-// Convert a string into a *_HUD_BOX_IDX constant
-int HUDRegionNameToIndex(char *name);
-
-class DCHUDRegion {
-public:
-	Box coords;
-	Box uv_erase_coords;
-	uvfloat4 erase_coords;
-	bool bLimitsComputed;
-};
-
-/*
- * This class stores the coordinates for each HUD Region : left radar, right radar, text
- * boxes, etc. It does not store the individual HUD elements within each HUD texture. For
- * that, look at DCElementSourceBox
- */
-class DCHUDRegions {
-public:
-	std::vector<DCHUDRegion> boxes;
-
-	DCHUDRegions();
-
-	void Clear() {
-		boxes.clear();
-	}
-
-	void ResetLimits() {
-		for (unsigned int i = 0; i < boxes.size(); i++) 
-			boxes[i].bLimitsComputed = false;
-	}
-};
-const int MAX_DC_REGIONS = 9;
-
-const int LEFT_RADAR_DC_ELEM_SRC_IDX = 0;
-const int RIGHT_RADAR_DC_ELEM_SRC_IDX = 1;
-const int LASER_RECHARGE_DC_ELEM_SRC_IDX = 2;
-const int SHIELD_RECHARGE_DC_ELEM_SRC_IDX = 3;
-const int ENGINE_RECHARGE_DC_ELEM_SRC_IDX = 4;
-const int BEAM_RECHARGE_DC_ELEM_SRC_IDX = 5;
-const int SHIELDS_DC_ELEM_SRC_IDX = 6;
-const int BEAM_DC_ELEM_SRC_IDX = 7;
-const int TARGET_COMP_DC_ELEM_SRC_IDX = 8;
-const int QUAD_LASERS_L_DC_ELEM_SRC_IDX = 9;
-const int QUAD_LASERS_R_DC_ELEM_SRC_IDX = 10;
-const int LEFT_MSG_DC_ELEM_SRC_IDX = 11;
-const int RIGHT_MSG_DC_ELEM_SRC_IDX = 12;
-const int SPEED_N_THROTTLE_DC_ELEM_SRC_IDX = 13;
-const int MISSILES_DC_ELEM_SRC_IDX = 14;
-const int NAME_TIME_DC_ELEM_SRC_IDX = 15;
-const int NUM_CRAFTS_DC_ELEM_SRC_IDX = 16;
-const int QUAD_LASERS_BOTH_DC_ELEM_SRC_IDX = 17;
-const int DUAL_LASERS_L_DC_ELEM_SRC_IDX = 18;
-const int DUAL_LASERS_R_DC_ELEM_SRC_IDX = 19;
-const int DUAL_LASERS_BOTH_DC_ELEM_SRC_IDX = 20;
-const int B_WING_LASERS_DC_ELEM_SRC_IDX = 21;
-const int SIX_LASERS_BOTH_DC_ELEM_SRC_IDX = 22;
-const int SIX_LASERS_L_DC_ELEM_SRC_IDX = 23;
-const int SIX_LASERS_R_DC_ELEM_SRC_IDX = 24;
-const int SHIELDS_FRONT_DC_ELEM_SRC_IDX = 25;
-const int SHIELDS_BACK_DC_ELEM_SRC_IDX = 26;
-const int KW_TEXT_CMD_DC_ELEM_SRC_IDX = 27;
-const int KW_TEXT_TOP_DC_ELEM_SRC_IDX = 28;
-const int KW_TEXT_RADIOSYS_DC_ELEM_SRC_IDX = 29;
-const int TEXT_RADIO_DC_ELEM_SRC_IDX = 30;
-const int TEXT_SYSTEM_DC_ELEM_SRC_IDX = 31;
-const int TEXT_CMD_DC_ELEM_SRC_IDX = 32;
-const int TARGETED_OBJ_NAME_SRC_IDX = 33;
-const int TARGETED_OBJ_SHD_SRC_IDX = 34;
-const int TARGETED_OBJ_HULL_SRC_IDX = 35;
-const int TARGETED_OBJ_CARGO_SRC_IDX = 36;
-const int TARGETED_OBJ_SYS_SRC_IDX = 37;
-const int TARGETED_OBJ_DIST_SRC_IDX = 38;
-const int TARGETED_OBJ_SUBCMP_SRC_IDX = 39;
-const int MAX_DC_SRC_ELEMENTS = 40;
-extern std::vector<const char *>g_DCElemSrcNames;
-// Convert a string into a *_DC_ELEM_SRC_IDX constant
-int DCSrcElemNameToIndex(char *name);
-
-class DCElemSrcBox {
-public:
-	Box uv_coords;
-	Box coords;
-	bool bComputed;
-
-	DCElemSrcBox() {
-		bComputed = false;
-	}
-};
-
-/*
- * Stores the uv_coords and pixel coords for each individual HUD element. Examples of HUD elems
- * are:
- * Laser recharge rate, Shield recharage rate, Radars, etc.
- */
-class DCElemSrcBoxes {
-public:
-	std::vector<DCElemSrcBox> src_boxes;
-
-	void Clear() {
-		src_boxes.clear();
-	}
-
-	DCElemSrcBoxes();
-	void Reset() {
-		for (unsigned int i = 0; i < src_boxes.size(); i++)
-			src_boxes[i].bComputed = false;
-	}
-};
+#include "effects.h"
+#include "dynamic_cockpit.h"
+#include "active_cockpit.h"
 
 enum RenderMainColorKeyType
 {
@@ -186,17 +32,6 @@ typedef struct HeadPosStruct {
 } HeadPos;
 /*****************/
 
-/* 2D Constant Buffers */
-typedef struct MainShadersCBStruct {
-	float scale, aspectRatio, parallax, brightness;
-	float use_3D, inv_scale, unused1, unused2;
-} MainShadersCBuffer;
-
-typedef struct BarrelPixelShaderCBStruct {
-	float k1, k2, k3;
-	int unused;
-} BarrelPixelShaderCBuffer;
-
 #define BACKBUFFER_FORMAT DXGI_FORMAT_B8G8R8A8_UNORM
 //#define BLOOM_BUFFER_FORMAT DXGI_FORMAT_B8G8R8A8_UNORM
 #define BLOOM_BUFFER_FORMAT DXGI_FORMAT_R16G16B16A16_FLOAT
@@ -206,229 +41,7 @@ typedef struct BarrelPixelShaderCBStruct {
 #define AO_MASK_FORMAT DXGI_FORMAT_B8G8R8A8_UNORM
 #define HDR_FORMAT DXGI_FORMAT_R16G16B16A16_FLOAT
 
-typedef struct BloomConfigStruct {
-	float fSaturationStrength, fCockpitStrength, fEngineGlowStrength, fSparksStrength;
-	float fLightMapsStrength, fLasersStrength, fHyperStreakStrength, fHyperTunnelStrength;
-	float fTurboLasersStrength, fLensFlareStrength, fExplosionsStrength, fSunsStrength;
-	float fCockpitSparksStrength, fMissileStrength, fSkydomeLightStrength, fBracketStrength;
-	float uvStepSize1, uvStepSize2;
-	int iNumPasses;
-} BloomConfig;
 
-typedef struct BloomPixelShaderCBStruct {
-	float pixelSizeX, pixelSizeY, general_bloom_strength, amplifyFactor;
-	// 16 bytes
-	float bloomStrength, uvStepSize, saturationStrength;
-	int unused1;
-	// 32 bytes
-	int unused2;
-	float unused3, depth_weight;
-	int debug;
-	// 48 bytes
-} BloomPixelShaderCBuffer;
-
-typedef struct SSAOPixelShaderCBStruct {
-	float screenSizeX, screenSizeY, indirect_intensity, bias;
-	// 16 bytes
-	float intensity, near_sample_radius, black_level;
-	int samples;
-	// 32 bytes
-	int z_division;
-	float bentNormalInit, max_dist, power;
-	// 48 bytes
-	int debug;
-	float moire_offset, amplifyFactor;
-	int fn_enable;
-	// 64 bytes
-	float fn_max_xymult, fn_scale, fn_sharpness, nm_intensity_near;
-	// 80 bytes
-	float far_sample_radius, nm_intensity_far, ssao_unused0, amplifyFactor2;
-	// 96 bytes
-	float x0, y0, x1, y1; // Viewport limits in uv space
-	// 112 bytes
-	float enable_dist_fade, ssaops_unused1, ssaops_unused2, shadow_epsilon;
-	// 128 bytes
-	float ssaops_unused3, shadow_step_size, shadow_steps, aspect_ratio;
-	// 144 bytes
-	float vpScale[4];
-	// 160 bytes
-	int shadow_enable;
-	float shadow_k, Bz_mult, moire_scale;
-	// 176 bytes
-} SSAOPixelShaderCBuffer;
-
-typedef struct ShadertoyCBStruct {
-	float iTime, twirl, bloom_strength, srand;
-	// 16 bytes
-	float iResolution[2];
-	// 0: Non-VR, 1: DirectSBS, 2: SteamVR. 
-	// Set to 1 when the Hyperspace Effect is in the HS_POST_HYPER_EXIT_ST state
-	// Used in the SunShader too.
-	int VRmode; 
-	float y_center;
-	// 32 bytes
-	float x0, y0, x1, y1; // Limits in uv-coords of the viewport
-	// 48 bytes
-	Matrix4 viewMat; // The view rotation matrix
-	// 4*4 = 16 elements, 16 * 4 = 64 bytes
-	// 48 + 64 = 112 bytes
-	int bDisneyStyle; // Enables the flare when jumping into hyperspace and other details
-	int hyperspace_phase; // 1 = HYPER_ENTRY, 2 = HYPER_TUNNEL, 3 = HYPER_EXIT, 4 = POST_HYPER_EXIT (same as HypespacePhaseEnum)
-	float tunnel_speed, FOVscale;
-	// 128 bytes
-	int SunFlareCount;
-	float flare_intensity;
-	float preserveAspectRatioComp[2]; // Used to compensate for the distortion introduced when PreserveAspectRatio = 0 in DDraw.cfg
-	// 144 bytes
-	//float SunX, SunY, SunZ, flare_intensity;
-	float4 SunCoords[MAX_SUN_FLARES];
-	// 208 bytes
-	float4 SunColor[MAX_SUN_FLARES];
-	// 272 bytes
-} ShadertoyCBuffer;
-
-// Let's make this Constant Buffer the same size as the ShadertoyCBuffer
-// so that we can reuse the same CB slot -- after all, we can't manipulate
-// anything while travelling through hyperspace anyway...
-typedef struct LaserPointerCBStruct {
-	int TriggerState; // 0 = Not pressed, 1 = Pressed
-	float FOVscale, iResolution[2];
-	// 16 bytes
-	float x0, y0, x1, y1; // Limits in uv-coords of the viewport
-	// 32 bytes
-	float contOrigin[2], intersection[2];
-	// 48 bytes
-	int bContOrigin, bIntersection, bHoveringOnActiveElem;
-	int DirectSBSEye;
-	// 64 bytes
-	float v0[2], v1[2]; // DEBUG
-	// 80 bytes
-	float v2[2], uv[2]; // DEBUG
-	// 96
-	int bDebugMode;
-	float cursor_radius, lp_aspect_ratio[2];
-	// 112 bytes
-} LaserPointerCBuffer;
-
-/* 3D Constant Buffers */
-typedef struct VertexShaderCBStruct {
-	float viewportScale[4];
-	// 16 bytes
-	float aspect_ratio;
-	uint32_t apply_uv_comp;
-	float z_override, sz_override;
-	// 32 bytes
-	float mult_z_override, bPreventTransform, bFullTransform, scale_override;
-	// 48 bytes
-	//float vsunused0, vsunused1, vsunused2, vsunused3;
-	// 64 bytes
-} VertexShaderCBuffer;
-
-typedef struct VertexShaderMatrixCBStruct {
-	Matrix4 projEye;
-	Matrix4 viewMat;
-	Matrix4 fullViewMat;
-} VertexShaderMatrixCB;
-
-typedef struct PSShadingSystemCBStruct {
-	float3 MainLight;
-	uint32_t LightCount;
-	// 16 bytes
-	float4 MainColor;
-	// 32 bytes
-	float4 LightVector[MAX_XWA_LIGHTS];
-	// 32+128 = 160 bytes
-	float4 LightColor[MAX_XWA_LIGHTS];
-	// 160+128 = 288 bytes
-	float spec_intensity, glossiness, spec_bloom_intensity, bloom_glossiness_mult;
-	// 304 bytes
-	float saturation_boost, lightness_boost, ssdo_enabled;
-	uint32_t ss_debug;
-	// 320 bytes
-	float sso_disable, sqr_attenuation, laser_light_intensity;
-	uint32_t num_lasers;
-	// 336 bytes
-	float4 LightPoint[MAX_CB_POINT_LIGHTS];
-	// 8 * 16 = 128
-	// 464 bytes
-	float4 LightPointColor[MAX_CB_POINT_LIGHTS];
-	// 8 * 16 = 128
-	// 592 bytes
-	float ambient, headlights_angle_cos, HDR_white_point;
-	uint32_t HDREnabled;
-	// 608 bytes
-} PSShadingSystemCB;
-
-// See PixelShaderTextureCommon.h for an explanation of these settings
-typedef struct PixelShaderCBStruct {
-	float brightness;			// Used to control the brightness of some elements -- mostly for ReShade compatibility
-	uint32_t DynCockpitSlots;
-	uint32_t bUseCoverTexture;
-	uint32_t bInHyperspace;
-	// 16 bytes
-	
-	uint32_t bIsLaser;
-	uint32_t bIsLightTexture;
-	uint32_t bIsEngineGlow;
-	uint32_t ps_unused1;
-	// 32 bytes
-
-	float fBloomStrength;
-	float fPosNormalAlpha;
-	float fSSAOMaskVal;
-	float fSSAOAlphaMult;
-	// 48 bytes
-
-	uint32_t bIsShadeless;
-	float fGlossiness, fSpecInt, fNMIntensity;
-	// 64 bytes
-
-	float fSpecVal, fDisableDiffuse;
-	uint32_t special_control;
-	float fAmbient;
-	// 80 bytes
-} PixelShaderCBuffer;
-
-// Pixel Shader constant buffer for the Dynamic Cockpit
-typedef struct DCPixelShaderCBStruct {
-	uvfloat4 src[MAX_DC_COORDS_PER_TEXTURE];
-	// 4 * MAX_DC_COORDS * 4 = 192
-	uvfloat4 dst[MAX_DC_COORDS_PER_TEXTURE];
-	// 4 * MAX_DC_COORDS * 4 = 192
-	// 384 bytes thus far
-	uint32_t bgColor[MAX_DC_COORDS_PER_TEXTURE]; // 32-bit Background colors
-	// 4 * MAX_DC_COORDS = 48
-	// 432 bytes thus far
-
-	float ct_brightness, dc_brightness;
-	uint32_t noisy_holo; // If set to 1, the hologram shader will be noisy!
-	float transparent; // If set to 1, the background will be transparent
-	// 448 bytes
-} DCPixelShaderCBuffer;
-
-// Vertex Shader constant buffer used in ShadowMapVS.hlsl, register b5
-typedef struct ShadowMapVertexShaderMatrixCBStruct {
-	Matrix4 Camera;
-	Matrix4 lightWorldMatrix[MAX_XWA_LIGHTS];
-	// 128 bytes
-
-	uint32_t sm_enabled, sm_debug;
-	float sm_light_size, sm_blocker_radius;
-
-	float sm_aspect_ratio, sm_bias, sm_unused, sm_pcss_radius;
-
-	Vector3 POV;
-	float sm_resolution;
-
-	int light_index;
-	float sm_FOVscale, sm_y_center, sm_z_factor;
-
-	uint32_t sm_PCSS_enabled, sm_pcss_samples, sm_hardware_pcf, sm_VR_mode;
-
-	float sm_black_levels[MAX_XWA_LIGHTS]; // 8 levels: 2 16-byte rows
-	float OBJrange[MAX_XWA_LIGHTS]; // 8 ranges: 2 16-byte rows
-	float OBJminZ[MAX_XWA_LIGHTS]; // 8 values: 2 16-byte rows
-} ShadowMapVertexShaderMatrixCB;
 
 // Holds the current 3D reconstruction constants, register b6
 typedef struct MetricReconstructionCBStruct {
@@ -444,85 +57,9 @@ typedef struct MetricReconstructionCBStruct {
 	float mv_vr_vertexbuf_aspect_ratio_comp[2]; // This is used to render the HUD
 } MetricReconstructionCB;
 
-typedef struct uv_coords_src_dst_struct {
-	int src_slot[MAX_DC_COORDS_PER_TEXTURE]; // This src slot references one of the pre-defined DC internal areas
-	uvfloat4 dst[MAX_DC_COORDS_PER_TEXTURE];
-	uint32_t uBGColor[MAX_DC_COORDS_PER_TEXTURE];
-	uint32_t uHGColor[MAX_DC_COORDS_PER_TEXTURE];
-	uint32_t uWHColor[MAX_DC_COORDS_PER_TEXTURE];
-	int numCoords;
-} uv_src_dst_coords;
 
-typedef struct uv_coords_struct {
-	uvfloat4 src[MAX_DC_COORDS_PER_TEXTURE];
-	int numCoords;
-} uv_coords;
-
-const int MAX_TEXTURE_NAME = 128;
-typedef struct dc_element_struct {
-	uv_src_dst_coords coords;
-	int erase_slots[MAX_DC_COORDS_PER_TEXTURE];
-	int num_erase_slots;
-	char name[MAX_TEXTURE_NAME];
-	char coverTextureName[MAX_TEXTURE_NAME];
-	//ComPtr<ID3D11ShaderResourceView> coverTexture = nullptr;
-	//ID3D11ShaderResourceView *coverTexture = NULL;
-	bool bActive, bNameHasBeenTested, bHologram, bNoisyHolo, bTransparent;
-} dc_element;
-
-typedef struct move_region_coords_struct {
-	int region_slot[MAX_HUD_BOXES];
-	uvfloat4 dst[MAX_HUD_BOXES];
-	int numCoords;
-} move_region_coords;
-
-// ACTIVE COCKPIT
-#define MAX_AC_COORDS_PER_TEXTURE 64
-#define MAX_AC_TEXTURES_PER_COCKPIT 16
-#define MAX_AC_ACTION_LEN 8 // WORDs (scan codes) used to specify an action
-#define AC_HOLOGRAM_FAKE_VK_CODE 0x01 // Internal AC code to toggle the holograms
-typedef struct ac_uv_coords_struct {
-	uvfloat4 area[MAX_AC_COORDS_PER_TEXTURE];
-	WORD action[MAX_AC_COORDS_PER_TEXTURE][MAX_AC_ACTION_LEN]; // List of scan codes
-	char action_name[MAX_AC_COORDS_PER_TEXTURE][16]; // For debug purposes only, remove later
-	int numCoords;
-} ac_uv_coords;
-
-typedef struct ac_element_struct {
-	ac_uv_coords coords;
-	//int idx; // "Back pointer" into the g_ACElements array
-	char name[MAX_TEXTURE_NAME];
-	bool bActive, bNameHasBeenTested;
-	short width, height; // DEBUG, remove later
-} ac_element;
-
-// SSAO Type
-typedef enum {
-	SSO_AMBIENT,
-	SSO_DIRECTIONAL,
-	SSO_BENT_NORMALS,
-	SSO_DEFERRED, // New Shading Model
-} SSAOTypeEnum;
-
-// In order to blend the background with the hyperspace effect when exiting, we need to extend the
-// effect for a few more frames. To do that, we need to track the current state of the effect and
-// that's why we need a small state machine:
-enum HyperspacePhaseEnum {
-	HS_INIT_ST = 0,				// Initial state, we're not even in Hyperspace
-	HS_HYPER_ENTER_ST = 1,		// We're entering hyperspace
-	HS_HYPER_TUNNEL_ST = 2,		// Traveling through the blue Hyperspace tunnel
-	HS_HYPER_EXIT_ST = 3,		// HyperExit streaks are being rendered
-	HS_POST_HYPER_EXIT_ST = 4   // HyperExit streaks have finished rendering; but now we're blending with the backround
-};
-const int MAX_POST_HYPER_EXIT_FRAMES = 10; // I had 20 here up to version 1.1.1. Making this smaller makes the zoom faster
 
 // General types and globals
-typedef enum {
-	TRACKER_NONE,
-	TRACKER_FREEPIE,
-	TRACKER_STEAMVR,
-	TRACKER_TRACKIR
-} TrackerType;
 
 struct MaterialStruct;
 extern struct MaterialStruct g_DefaultGlobalMaterial;
@@ -622,24 +159,6 @@ typedef struct AuxTextureDataStruct {
 } AuxTextureData;
 */
 
-// For shadow mapping, maybe we'd like to distinguish between sun-lights and
-// planet/background-lights. We'll use this struct to tag lights and fade
-// those lights which aren't suns
-class XWALightInfo {
-public:
-	bool bTagged, bIsSun;
-
-public:
-	XWALightInfo() {
-		this->Reset();
-	}
-
-	void Reset() {
-		this->bTagged = false;
-		this->bIsSun = true;
-	}
-};
-
 // Text Rendering
 // Font indices that can be used with the PrimarySurface::AddText() methods (and others) below
 #define FONT_MEDIUM_IDX 0
@@ -738,87 +257,7 @@ public:
 	void insert(Vector3 P, Vector3 col);
 };
 
-#define MAX_SPEED_PARTICLES 256
-extern Vector4 g_SpeedParticles[MAX_SPEED_PARTICLES];
 
-#define SHADOW_MAP_SIZE 1024
-
-class ShadowMappingData {
-public:
-	bool bEnabled;
-	bool bAnisotropicMapScale;
-	bool bAllLightsTagged;
-	bool bMultipleSuns;
-	bool bUseShadowOBJ; // This should be set to true when the Shadow OBJ is loaded
-	bool bOBJrange_override;
-	float fOBJrange_override_value;
-	int ShadowMapSize;
-	D3D11_VIEWPORT ViewPort;
-	int NumVertices, NumIndices; // This should be set when the Shadow OBJ is loaded
-	float black_level;
-	float POV_XY_FACTOR;
-	float POV_Z_FACTOR;
-	float FOVDistScale;
-	float sw_pcf_bias;
-	float hw_pcf_bias;
-	//float XWA_LIGHT_Y_CONV_SCALE;
-	float shadow_map_mult_x;
-	float shadow_map_mult_y;
-	float shadow_map_mult_z;
-
-	int DepthBias;
-	float DepthBiasClamp;
-	float SlopeScaledDepthBias;
-
-	ShadowMappingData() {
-		this->bEnabled = false;
-		this->bAnisotropicMapScale = true;
-		this->bAllLightsTagged = false;
-		this->bMultipleSuns = false;
-		this->bUseShadowOBJ = false;
-		this->NumVertices = 0;
-		this->NumIndices = 0;
-		this->ShadowMapSize   = SHADOW_MAP_SIZE;
-		// Initialize the Viewport
-		this->ViewPort.TopLeftX = 0.0f;
-		this->ViewPort.TopLeftY = 0.0f;
-		this->ViewPort.Height   = (float)this->ShadowMapSize;
-		this->ViewPort.Width    = (float)this->ShadowMapSize;
-		this->ViewPort.MinDepth = D3D11_MIN_DEPTH;
-		this->ViewPort.MaxDepth = D3D11_MAX_DEPTH;
-		this->black_level = 0.2f;
-		this->POV_XY_FACTOR = 24.974f;
-		this->POV_Z_FACTOR = 25.0f;
-		this->bAnisotropicMapScale = true;
-		this->bOBJrange_override = false;
-		this->fOBJrange_override_value = 5.0f;
-		//this->FOVDistScale = 624.525f;
-		this->FOVDistScale = 620.0f; // This one seems a bit better
-		this->sw_pcf_bias = -0.03f;
-		this->hw_pcf_bias = -0.03f;
-		// The following scale factor is used when tagging lights (associating XWA lights
-		// with sun textures). I don't have a good explanation for this value; but
-		// it's used to compensate the Y coordinate so that the light and the centroid of
-		// the sun texture line up better. I'll investigate this in detail later.
-		//this->XWA_LIGHT_Y_CONV_SCALE = -62.5f;
-
-		this->shadow_map_mult_x =  1.0f;
-		this->shadow_map_mult_y =  1.0f;
-		this->shadow_map_mult_z = -1.0f;
-
-		this->DepthBias = 0;
-		this->DepthBiasClamp = 0.0f;
-		this->SlopeScaledDepthBias = 0.0f;
-	}
-
-	void SetSize(int Width, int Height) {
-		this->ShadowMapSize = Width;
-		this->ViewPort.Width = (float)ShadowMapSize;
-		this->ViewPort.Height = (float)ShadowMapSize;
-	}
-};
-
-extern ShadowMappingData g_ShadowMapping;
 
 class DeviceResources
 {
