+++ impl11/ddraw/DirectSBS.h
@@ -1,4 +1,5 @@
 #pragma once
 
-
 bool InitDirectSBS();
+bool InitDirectSBS();
+bool ShutDownDirectSBS();
