+++ impl11/ddraw/effects.h
@@ -1,0 +1,290 @@
+#pragma once
+
+#include "Matrices.h"
+#include "../shaders/material_defs.h"
+#include "../shaders/shader_common.h"
+#include <vector>
+
+typedef struct Box_struct {
+	float x0, y0, x1, y1;
+} Box;
+
+typedef struct uvfloat4_struct {
+	float x0, y0, x1, y1;
+} uvfloat4;
+
+typedef struct float3_struct {
+	float x, y, z;
+} float3;
+
+typedef struct float4_struct {
+	float x, y, z, w;
+} float4;
+
+// SSAO Type
+typedef enum {
+	SSO_AMBIENT,
+	SSO_DIRECTIONAL,
+	SSO_BENT_NORMALS,
+	SSO_DEFERRED, // New Shading Model
+} SSAOTypeEnum;
+
+// In order to blend the background with the hyperspace effect when exiting, we need to extend the
+// effect for a few more frames. To do that, we need to track the current state of the effect and
+// that's why we need a small state machine:
+enum HyperspacePhaseEnum {
+	HS_INIT_ST = 0,				// Initial state, we're not even in Hyperspace
+	HS_HYPER_ENTER_ST = 1,		// We're entering hyperspace
+	HS_HYPER_TUNNEL_ST = 2,		// Traveling through the blue Hyperspace tunnel
+	HS_HYPER_EXIT_ST = 3,		// HyperExit streaks are being rendered
+	HS_POST_HYPER_EXIT_ST = 4   // HyperExit streaks have finished rendering; but now we're blending with the backround
+};
+const int MAX_POST_HYPER_EXIT_FRAMES = 10; // I had 20 here up to version 1.1.1. Making this smaller makes the zoom faster
+
+typedef struct BloomConfigStruct {
+	float fSaturationStrength, fCockpitStrength, fEngineGlowStrength, fSparksStrength;
+	float fLightMapsStrength, fLasersStrength, fHyperStreakStrength, fHyperTunnelStrength;
+	float fTurboLasersStrength, fLensFlareStrength, fExplosionsStrength, fSunsStrength;
+	float fCockpitSparksStrength, fMissileStrength, fSkydomeLightStrength, fBracketStrength;
+	float uvStepSize1, uvStepSize2;
+	int iNumPasses;
+} BloomConfig;
+
+typedef struct BloomPixelShaderCBStruct {
+	float pixelSizeX, pixelSizeY, general_bloom_strength, amplifyFactor;
+	// 16 bytes
+	float bloomStrength, uvStepSize, saturationStrength;
+	int unused1;
+	// 32 bytes
+	int unused2;
+	float unused3, depth_weight;
+	int debug;
+	// 48 bytes
+} BloomPixelShaderCBuffer;
+
+typedef struct SSAOPixelShaderCBStruct {
+	float screenSizeX, screenSizeY, indirect_intensity, bias;
+	// 16 bytes
+	float intensity, near_sample_radius, black_level;
+	int samples;
+	// 32 bytes
+	int z_division;
+	float bentNormalInit, max_dist, power;
+	// 48 bytes
+	int debug;
+	float moire_offset, amplifyFactor;
+	int fn_enable;
+	// 64 bytes
+	float fn_max_xymult, fn_scale, fn_sharpness, nm_intensity_near;
+	// 80 bytes
+	float far_sample_radius, nm_intensity_far, ssao_unused0, amplifyFactor2;
+	// 96 bytes
+	float x0, y0, x1, y1; // Viewport limits in uv space
+	// 112 bytes
+	float enable_dist_fade, ssaops_unused1, ssaops_unused2, shadow_epsilon;
+	// 128 bytes
+	float ssaops_unused3, shadow_step_size, shadow_steps, aspect_ratio;
+	// 144 bytes
+	float vpScale[4];
+	// 160 bytes
+	int shadow_enable;
+	float shadow_k, Bz_mult, moire_scale;
+	// 176 bytes
+} SSAOPixelShaderCBuffer;
+
+typedef struct ShadertoyCBStruct {
+	float iTime, twirl, bloom_strength, srand;
+	// 16 bytes
+	float iResolution[2];
+	// 0: Non-VR, 1: DirectSBS, 2: SteamVR. 
+	// Set to 1 when the Hyperspace Effect is in the HS_POST_HYPER_EXIT_ST state
+	// Used in the SunShader too.
+	int VRmode;
+	float y_center;
+	// 32 bytes
+	float x0, y0, x1, y1; // Limits in uv-coords of the viewport
+	// 48 bytes
+	Matrix4 viewMat; // The view rotation matrix
+	// 4*4 = 16 elements, 16 * 4 = 64 bytes
+	// 48 + 64 = 112 bytes
+	int bDisneyStyle; // Enables the flare when jumping into hyperspace and other details
+	int hyperspace_phase; // 1 = HYPER_ENTRY, 2 = HYPER_TUNNEL, 3 = HYPER_EXIT, 4 = POST_HYPER_EXIT (same as HypespacePhaseEnum)
+	float tunnel_speed, FOVscale;
+	// 128 bytes
+	int SunFlareCount;
+	float flare_intensity;
+	float preserveAspectRatioComp[2]; // Used to compensate for the distortion introduced when PreserveAspectRatio = 0 in DDraw.cfg
+	// 144 bytes
+	//float SunX, SunY, SunZ, flare_intensity;
+	float4 SunCoords[MAX_SUN_FLARES];
+	// 208 bytes
+	float4 SunColor[MAX_SUN_FLARES];
+	// 272 bytes
+} ShadertoyCBuffer;
+
+// Let's make this Constant Buffer the same size as the ShadertoyCBuffer
+// so that we can reuse the same CB slot -- after all, we can't manipulate
+// anything while travelling through hyperspace anyway...
+typedef struct LaserPointerCBStruct {
+	int TriggerState; // 0 = Not pressed, 1 = Pressed
+	float FOVscale, iResolution[2];
+	// 16 bytes
+	float x0, y0, x1, y1; // Limits in uv-coords of the viewport
+	// 32 bytes
+	float contOrigin[2], intersection[2];
+	// 48 bytes
+	int bContOrigin, bIntersection, bHoveringOnActiveElem;
+	int DirectSBSEye;
+	// 64 bytes
+	float v0[2], v1[2]; // DEBUG
+	// 80 bytes
+	float v2[2], uv[2]; // DEBUG
+	// 96
+	int bDebugMode;
+	float cursor_radius, lp_aspect_ratio[2];
+	// 112 bytes
+} LaserPointerCBuffer;
+
+/* 3D Constant Buffers */
+typedef struct VertexShaderCBStruct {
+	float viewportScale[4];
+	// 16 bytes
+	float aspect_ratio;
+	uint32_t apply_uv_comp;
+	float z_override, sz_override;
+	// 32 bytes
+	float mult_z_override, bPreventTransform, bFullTransform, scale_override;
+	// 48 bytes
+	//float vsunused0, vsunused1, vsunused2, vsunused3;
+	// 64 bytes
+} VertexShaderCBuffer;
+
+typedef struct VertexShaderMatrixCBStruct {
+	Matrix4 projEye;
+	Matrix4 viewMat;
+	Matrix4 fullViewMat;
+} VertexShaderMatrixCB;
+
+typedef struct PSShadingSystemCBStruct {
+	float3 MainLight;
+	uint32_t LightCount;
+	// 16 bytes
+	float4 MainColor;
+	// 32 bytes
+	float4 LightVector[MAX_XWA_LIGHTS];
+	// 32+128 = 160 bytes
+	float4 LightColor[MAX_XWA_LIGHTS];
+	// 160+128 = 288 bytes
+	float spec_intensity, glossiness, spec_bloom_intensity, bloom_glossiness_mult;
+	// 304 bytes
+	float saturation_boost, lightness_boost, ssdo_enabled;
+	uint32_t ss_debug;
+	// 320 bytes
+	float sso_disable, sqr_attenuation, laser_light_intensity;
+	uint32_t num_lasers;
+	// 336 bytes
+	float4 LightPoint[MAX_CB_POINT_LIGHTS];
+	// 8 * 16 = 128
+	// 464 bytes
+	float4 LightPointColor[MAX_CB_POINT_LIGHTS];
+	// 8 * 16 = 128
+	// 592 bytes
+	float ambient, headlights_angle_cos, HDR_white_point;
+	uint32_t HDREnabled;
+	// 608 bytes
+} PSShadingSystemCB;
+
+// See PixelShaderTextureCommon.h for an explanation of these settings
+typedef struct PixelShaderCBStruct {
+	float brightness;			// Used to control the brightness of some elements -- mostly for ReShade compatibility
+	uint32_t DynCockpitSlots;
+	uint32_t bUseCoverTexture;
+	uint32_t bInHyperspace;
+	// 16 bytes
+
+	uint32_t bIsLaser;
+	uint32_t bIsLightTexture;
+	uint32_t bIsEngineGlow;
+	uint32_t ps_unused1;
+	// 32 bytes
+
+	float fBloomStrength;
+	float fPosNormalAlpha;
+	float fSSAOMaskVal;
+	float fSSAOAlphaMult;
+	// 48 bytes
+
+	uint32_t bIsShadeless;
+	float fGlossiness, fSpecInt, fNMIntensity;
+	// 64 bytes
+
+	float fSpecVal, fDisableDiffuse;
+	uint32_t special_control;
+	float fAmbient;
+	// 80 bytes
+} PixelShaderCBuffer;
+
+// Pixel Shader constant buffer for the Dynamic Cockpit
+typedef struct DCPixelShaderCBStruct {
+	uvfloat4 src[MAX_DC_COORDS_PER_TEXTURE];
+	// 4 * MAX_DC_COORDS * 4 = 192
+	uvfloat4 dst[MAX_DC_COORDS_PER_TEXTURE];
+	// 4 * MAX_DC_COORDS * 4 = 192
+	// 384 bytes thus far
+	uint32_t bgColor[MAX_DC_COORDS_PER_TEXTURE]; // 32-bit Background colors
+	// 4 * MAX_DC_COORDS = 48
+	// 432 bytes thus far
+
+	float ct_brightness, dc_brightness;
+	uint32_t noisy_holo; // If set to 1, the hologram shader will be noisy!
+	float transparent; // If set to 1, the background will be transparent
+	// 448 bytes
+} DCPixelShaderCBuffer;
+
+// Vertex Shader constant buffer used in ShadowMapVS.hlsl, register b5
+typedef struct ShadowMapVertexShaderMatrixCBStruct {
+	Matrix4 Camera;
+	Matrix4 lightWorldMatrix[MAX_XWA_LIGHTS];
+	// 128 bytes
+
+	uint32_t sm_enabled, sm_debug;
+	float sm_light_size, sm_blocker_radius;
+
+	float sm_aspect_ratio, sm_bias, sm_unused, sm_pcss_radius;
+
+	Vector3 POV;
+	float sm_resolution;
+
+	int light_index;
+	float sm_FOVscale, sm_y_center, sm_z_factor;
+
+	uint32_t sm_PCSS_enabled, sm_pcss_samples, sm_hardware_pcf, sm_VR_mode;
+
+	float sm_black_levels[MAX_XWA_LIGHTS]; // 8 levels: 2 16-byte rows
+	float OBJrange[MAX_XWA_LIGHTS]; // 8 ranges: 2 16-byte rows
+	float OBJminZ[MAX_XWA_LIGHTS]; // 8 values: 2 16-byte rows
+} ShadowMapVertexShaderMatrixCB;
+
+/* 2D Constant Buffers */
+typedef struct MainShadersCBStruct {
+	float scale, aspectRatio, parallax, brightness;
+	float use_3D, inv_scale, unused1, unused2;
+} MainShadersCBuffer;
+
+typedef struct BarrelPixelShaderCBStruct {
+	float k1, k2, k3;
+	int unused;
+} BarrelPixelShaderCBuffer;
+
+#define MAX_SPEED_PARTICLES 256
+extern Vector4 g_SpeedParticles[MAX_SPEED_PARTICLES];
+
+// Constant Buffers
+extern BloomPixelShaderCBuffer		g_BloomPSCBuffer;
+extern SSAOPixelShaderCBuffer		g_SSAO_PSCBuffer;
+extern PSShadingSystemCB			g_ShadingSys_PSBuffer;
+extern ShadertoyCBuffer		g_ShadertoyBuffer;
+extern LaserPointerCBuffer	g_LaserPointerBuffer;
+
+extern bool g_b3DSunPresent, g_b3DSkydomePresent;
+
